refactoring, prop types
