+++ src/components/page-content/page-content.jsx
@@ -1,19 +1,12 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 import GenresList from './genres-list';
 import ShowMoreButton from './show-more-button';
 import FilmsList from './films-list';
 import {pageContentPropTypes} from '../../prop-types/prop-types';
 import {connect} from 'react-redux';
-import {fetchFilms} from '../../store/api-actions';
 
 const PageContent = (props) => {
-  const {films, countShowingFilms, filmListByGenre, isDataLoaded, loadFilms} = props;
-
-  useEffect(() => {
-    if(!isDataLoaded) {
-      loadFilms();
-    }
-  }, [isDataLoaded]);
+  const {films, countShowingFilms, filmListByGenre} = props;
 
   return (
     <React.Fragment>
@@ -55,13 +48,7 @@ const mapStateToProps = (state) => ({
   films: state.films,
   filmListByGenre: state.filmListByGenre,
   countShowingFilms: state.countShowingFilms,
-  isDataLoaded: state.isDataLoaded,
 });
 
-const mapDispatchToProps = (dispatch) => ({
-  loadFilms() {
-    dispatch(fetchFilms())
-  },
-});
 
-export default connect(mapStateToProps, mapDispatchToProps)(PageContent);
+export default connect(mapStateToProps, null)(PageContent);
